Feat: 로또구입 분리
