+++ src/main.js
@@ -3,11 +3,10 @@ import './css/reset.css';
 import './css/lotto.css';
 
 import './css/modal.css';
+import { lottos, purchaseEvent } from './lotto';
 
 
 const $purchaseButton = document.getElementsByClassName("purchase-button")[0];
-const $purchasemoney=document.getElementsByClassName("purchase-money")[0];
-const $lottolist= document.getElementsByClassName("lotto-tickets-container")[0];
 const $resultbutton=document.getElementsByClassName("result")[0];
 const $modal=document.getElementsByClassName("modal")[0];
 const $modalclosex=document.getElementsByClassName("modal-close-x")[0];
@@ -19,7 +18,6 @@ const $profit = document.querySelector('.profit');
 
 $modal.style.display = 'none';
 
-let lottos=[]                   //구입로또리스트
 let winMoneyCount=[0,0,0,0,0];  //당첨갯수리스트
 
 //클릭이벤트
@@ -28,51 +26,15 @@ $resultbutton.addEventListener("click",resultEvent);
 $modalclosex.addEventListener("click",toggleModal);
 $restartbutton.addEventListener("click",toggleModal);
 
-
-
-
-
-
-
-
-
-export function purchaseEvent(event){
-    event.preventDefault();
-    buyLottoHtmlInsert();
-}
-
 export function resultEvent(event){
     event.preventDefault();
-    Checkwinning();
-    modalHtmlInsert();
+    Checkwinning(lottos);
+    modalHtmlInsert(lottos);
     toggleModal();
 }
 
-//로또 구입
-function buyLottoHtmlInsert(){
-    lottos=[]
-    const boughtCount=Math.floor($purchasemoney.value/1000);
-    $lottolist.innerHTML = `<span class="lotto-bought-title">총  ${boughtCount}개를 구매하였습니다.</span>`
-    for(let count=0;count<boughtCount;count++){
-        $lottolist.innerHTML+=
-        `<div>🎟️ ${randomnumber()}</div>`
-    }
-}
-
-function randomnumber(){
-    const RandomRange=30;   //로또숫자범위
-    const lottoNumbers = [];//중복제거용
-    while(true){
-        const num= Math.floor(Math.random() * RandomRange)
-        if(!lottoNumbers.includes(num)){lottoNumbers.push(num);} 
-        if(lottoNumbers.length==6) break;
-    }
-    lottos.push(lottoNumbers)
-    return lottoNumbers.join(", ");
-}
-
 //당첨로또 체크
-function Checkwinning(){
+function Checkwinning(lottos){
     winMoneyCount=[0,0,0,0,0]
     const winning = Array.from($winninglottos).map(win => parseInt(win.value,10));
     const bonuswinning = $bonuswinninglotto.value;
@@ -94,7 +56,7 @@ function Checkwinning(){
 }
 
 //모달 작성
-function modalHtmlInsert(){
+function modalHtmlInsert(lottos){
     const winMoney=[5000,50000,1500000,30000000,2000000000]
     const winCount=["3개","4개","5개","5개+보너스볼","6개"]
     $resulttable.innerHTML=`
